+++ src/cljvm/speclj/cloverage.clj
@@ -1,21 +1,19 @@
 (ns speclj.cloverage
-  (:require [cloverage.coverage :as coverage]
-            [speclj.config :refer [*reporters* *runner*]]
-            [speclj.report.documentation]
-            [speclj.report.progress :as progress]
-            [speclj.results :as results]
-            [speclj.run.standard :as standard]
-            [speclj.running :refer [run-and-report]]))
+  (:require [cloverage.args :as args]
+            [cloverage.coverage :as coverage]
+            [speclj.cli :as cli]))
 
 ;; Assumes that cloverage is already in the classpath.
 
-(defmethod coverage/runner-fn :speclj [_opts]
-  ;(prn "_opts: " _opts)
-  (fn [nses]
-    (let [results (atom [])
-          runner (standard/->StandardRunner (atom []) results)
-          reporters [(progress/->ProgressReporter)]]
-      (binding [*runner* runner *reporters* reporters]
-        (apply require (map symbol nses))
-        (run-and-report runner reporters))
-      {:errors (results/fail-count @results)})))
+(defmethod coverage/runner-fn :speclj [opts]
+  (let [runner-opts (into {} (:runner-opts opts))]
+    (fn [_nses]
+      (let [result (apply cli/run (:args runner-opts))]
+        {:errors result}))))
+
+(defn -main [& args]
+  (let [[speclj-args cloverage-args] (split-with (complement #{"--"}) args)
+        speclj-opts    {:runner-opts {:args speclj-args}}
+        cloverage-args (concat ["-r" ":speclj"] (rest cloverage-args))
+        cloverage-opts (args/parse-args cloverage-args speclj-opts)]
+    (coverage/run-main cloverage-opts speclj-opts)))